Refactors RestMenu page with Material ui components, adds RestMenuStyles.js
