+++ src/components/MenuComponents/MenuCategory/MenuCategory.module.css
@@ -1,4 +1,4 @@
-.container {
+/* .container {
   width: 100%;
   display: flex;
   flex-direction: column;
@@ -14,4 +14,4 @@
   width: 100%;
   display: flex;
   flex-wrap: wrap;
-}
+} */
