+++ src/components/MenuComponents/MenuItems/MenuItems.module.css
@@ -1,5 +1,4 @@
-.container {
-    /* border: red solid 2px; */
+/* .container {
     display: flex;
     margin: 1.45rem;
     flex-direction: column;
@@ -119,4 +118,4 @@
 .optionTitle {
     font-size: 1.6rem;
     background-color: pink;
-}
+} */
