+++ src/components/MenuComponents/ViewMenuItem/ViewMenuItem.module.css
@@ -1,4 +1,4 @@
-.item {
+/* .item {
   display: flex;
   flex-wrap: wrap;
   flex-direction: column;
@@ -96,4 +96,4 @@
 
 .title img {
   max-width: 66%;
-}
+} */
