--- src/pages/RestMenu/RestMenu.module.css
@@ -1,5 +1,0 @@
-.add {
-  position: fixed;
-  bottom: 10%;
-  right: 10%;
-}
